Fix display of user-submitted search string
The original code for the search page failed to escape the search string parameter submitted by the user. This made it possible for an attack to inject Javascript code which would be executed in the user's browser. The vulnerability was detected by a security scan of the AccrualNet production web site performed in September 2014. The original code was also uppercasing the HTML markup surrounding the string, including the class name for the enclosing element for the count of search results. Both problems are corrected by this commit.

https://tracker.nci.nih.gov/browse/OCEDRUPAL-408